+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/components/estimators/AbstractTripRouterEstimator.java
@@ -2,10 +2,11 @@ package ch.ethz.matsim.discrete_mode_choice.components.estimators;
 
 import java.util.List;
 
-import org.matsim.api.core.v01.network.Network;
+import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.PlanElement;
-import org.matsim.core.router.LinkWrapperFacility;
+import org.matsim.core.router.ActivityWrapperFacility;
+import org.matsim.core.router.PlanRouter;
 import org.matsim.core.router.TripRouter;
 import org.matsim.facilities.ActivityFacilities;
 import org.matsim.facilities.Facility;
@@ -25,12 +26,10 @@ import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.TripCandi
  */
 public abstract class AbstractTripRouterEstimator implements TripEstimator {
 	private final TripRouter tripRouter;
-	private final Network network;
 	private final ActivityFacilities facilities;
 
-	public AbstractTripRouterEstimator(TripRouter tripRouter, Network network, ActivityFacilities facilities) {
+	public AbstractTripRouterEstimator(TripRouter tripRouter, ActivityFacilities facilities) {
 		this.tripRouter = tripRouter;
-		this.network = network;
 		this.facilities = facilities;
 	}
 
@@ -38,18 +37,8 @@ public abstract class AbstractTripRouterEstimator implements TripEstimator {
 	public final TripCandidate estimateTrip(Person person, String mode, DiscreteModeChoiceTrip trip,
 			List<TripCandidate> previousTrips) {
 		// I) Find the correct origin and destination facilities
-		Facility originFacility = facilities.getFacilities().get(trip.getOriginActivity().getFacilityId());
-
-		if (originFacility == null) {
-			originFacility = new LinkWrapperFacility(network.getLinks().get(trip.getOriginActivity().getLinkId()));
-		}
-
-		Facility destinationFacility = facilities.getFacilities().get(trip.getDestinationActivity().getFacilityId());
-
-		if (destinationFacility == null) {
-			destinationFacility = new LinkWrapperFacility(
-					network.getLinks().get(trip.getDestinationActivity().getLinkId()));
-		}
+		Facility originFacility = getFacilityForActivity(trip.getOriginActivity());
+		Facility destinationFacility = getFacilityForActivity(trip.getDestinationActivity());
 
 		// II) Perform the routing
 		List<? extends PlanElement> elements = tripRouter.calcRoute(mode, originFacility, destinationFacility,
@@ -57,6 +46,18 @@ public abstract class AbstractTripRouterEstimator implements TripEstimator {
 
 		// III) Perform utility estimation
 		return estimateTripCandidate(person, mode, trip, previousTrips, elements);
+	}
+
+	/**
+	 * This converts an activity to a facility. The logic is copied from
+	 * {@link PlanRouter#toFacility}.
+	 */
+	private Facility getFacilityForActivity(Activity activity) {
+		if (activity.getLinkId() == null && activity.getCoord() == null) {
+			return facilities.getFacilities().get(activity.getFacilityId());
+		}
+
+		return new ActivityWrapperFacility(activity);
 	}
 
 	/**